Improve utils
* Split them into 3 modules
* Switch argument order so inputs come first to be consistent
with rest of app
* Improve make_dir functionality to avoid deleting directories which
can be dangerous, and also add options to make code more concise
